refactor: common.type.ts를 도메인별 파일로 분리
- 안 쓰는 타입 삭제
