+++ src/api/history.api.ts
@@ -1,10 +1,10 @@
-import { GameHistoryResponse } from '@utils/types/common.type';
+import { GameHistoryResponse } from '@utils/types/history.type';
 
 import api from './index';
 
-
-export const getGameHistory = async  (by: string): Promise<GameHistoryResponse[]> => {
-    const { data } = await api.get(`/histories/me?by=${by}`);
-    return data.reverse();
-}
-
+export const getGameHistory = async (
+  by: string,
+): Promise<GameHistoryResponse[]> => {
+  const { data } = await api.get(`/histories/me?by=${by}`);
+  return data.reverse();
+};
