+++ src/utils/types/history.type.ts
@@ -1,0 +1,8 @@
+export interface GameHistoryResponse {
+  sessionId: number;
+  memberId: number;
+  score: number;
+  updatedAt: string;
+}
+
+export type HistoryViewType = 'DATE' | 'SESSION';
